Separate the RangedU8 type into its own module
The RangedU8 type needs to be separated into its own module such that
other modules are able to use it -- its previous module, location, would
be an inadequate home for the RangedU8 type if it is required by other
modules (as the RangedU8 type has no relation to the location module).

Also add ranged macro (similar to the nonzero macro) and use it to create
other ranged types (e.g., RangedU16, RangedI32) such that they can be
used if they are ever required -- hide the documentation for these types
unless they are used somewhere else in the code base to avoid cluttering
up the documentation.

Signed-off-by: Zachary Whiteley <z.n.whiteley@gmail.com>